Commit Changes

+++ models/staging/raw/stg_raw__ship.sql
@@ -21,5 +21,5 @@ renamed as (
 select orders_id,
         shipping_fee,
         logcost,
-        CAST(ship_cost AS INT64) AS ship_cost
+        CAST(ship_cost AS FLOAT64) AS ship_cost
         from renamed
